+++ src/pages/ClassManagement/LunchMenu/index.tsx
@@ -1,9 +1,8 @@
-import styled from 'styled-components';
-
 import useMe from '@Hooks/api/useMe';
 
 import Board from './Board';
 import RegisterShoolButton from './RegisterShoolButton/RegisterShoolButton';
+import * as S from './style';
 
 function LunchMenu() {
   const { data, isLoading } = useMe();
@@ -15,56 +14,24 @@ function LunchMenu() {
   } = data;
 
   const date = new Date();
-  console.log(date.getFullYear());
-  console.log(date.getMonth() + 1);
-  console.log(date.getDate());
 
   const standardDate = `${date.getFullYear()}년 ${
     date.getMonth() + 1
   }월 ${date.getDate()}일`;
 
   return (
-    <Layout>
+    <S.Layout>
       <div>
         {schoolName ? (
-          <SchoolName>{schoolName} 급식</SchoolName>
+          <S.SchoolName>{schoolName} 급식</S.SchoolName>
         ) : (
           <RegisterShoolButton />
         )}
       </div>
-      <StandardDate>{standardDate}</StandardDate>
+      <S.StandardDate>{standardDate}</S.StandardDate>
       <Board />
-    </Layout>
+    </S.Layout>
   );
 }
 
 export default LunchMenu;
-
-const Layout = styled.div`
-  display: grid;
-  grid-template-columns: 1fr auto;
-  grid-template-rows: auto 1fr;
-  align-items: center;
-  justify-items: start;
-  row-gap: 40px;
-
-  padding: 60px;
-
-  min-height: 100%;
-
-  color: ${(props) => props.theme.text};
-`;
-
-const SchoolName = styled.div`
-  padding: 8px 12px;
-  border-radius: 3px;
-
-  font-size: 2rem;
-
-  background-color: ${(props) => props.theme.background.primary};
-  color: ${(props) => props.theme.primaryText};
-`;
-
-const StandardDate = styled.div`
-  font-size: 2rem;
-`;
